moved report files to report folder in model
